# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend REST API for golang nuryanto2121",
        "title": "Base Playtopia",
        "contact": {
            "name": "Nuryanto",
            "url": "https://www.linkedin.com/in/nuryanto-1b2721156/",
            "email": "nuryantofattih@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/cms/outlets": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "GetList Outlets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "InitSearch",
                        "name": "initsearch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortField",
                        "name": "sortfield",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModelList"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Add Outlets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OutletForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
        "/v1/cms/outlets/lookup": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "GetDataLookUp Outlets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PerPage",
                        "name": "perpage",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "InitSearch",
                        "name": "initsearch",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "SortField",
                        "name": "sortfield",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseModelList"
                        }
                    }
                }
            }
        },
        "/v1/cms/outlets/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "GetById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Rubah Outlets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "req param #changes are possible to adjust the form of the registration form from frontend",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.OutletForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Delete Outlets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device Type",
                        "name": "Device-Type",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Version Apps",
                        "name": "Version",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language Apps",
                        "name": "Language",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "errors": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "models.AddOutletDetail": {
            "type": "object",
            "properties": {
                "outlet_id": {
                    "type": "string"
                },
                "outlet_price_weekday": {
                    "type": "number"
                },
                "outlet_price_weekend": {
                    "type": "number"
                },
                "product_id": {
                    "type": "string"
                }
            }
        },
...
        "models.OutletForm": {
            "type": "object",
            "properties": {
                "outlet_city": {
                    "type": "string"
                },
                "outlet_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AddOutletDetail"
                    }
                },
                "outlet_name": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseModelList": {
            "type": "object",
            "properties": {
                "data": {},
                "last_page": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}